Split buttons to related src and header
