+++ src/creo2urdf/src/main.cpp
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2006-2023 Istituto Italiano di Tecnologia (IIT)
+ * All rights reserved.
+ *
+ * This software may be modified and distributed under the terms of the
+ * BSD-3-Clause license. See the accompanying LICENSE file for details.
+ */
+#define _USE_MATH_DEFINES
+
+#include <creo2urdf/Creo2Urdf.h>
+#include <creo2urdf/Validator.h>
+
+/*====================================================================*\
+FUNCTION : user_initialize()
+PURPOSE  :
+\*====================================================================*/
+extern "C" int user_initialize(
+    int argc,
+    char* argv[],
+    char* version,
+    char* build,
+    wchar_t errbuf[80])
+{
+    auto session = pfcGetProESession();
+
+    auto cmd = session->UICreateCommand("Creo2Urdf", new Creo2UrdfActionListerner());
+    cmd->AddActionListener(new Creo2UrdfAccessListener()); // To be checked it is odd
+    cmd->Designate("ui.txt", "Run Creo2Urdf", "Run Creo2Urdf", "Run Creo2Urdf");
+
+    auto cmd_validate = session->UICreateCommand("Validator", new ValidatorListener());
+    cmd_validate->AddActionListener(new ValidatorAccessListener()); // To be checked it is odd
+    cmd_validate->Designate("ui.txt", "Run Validation", "Run Validation", "Run Validation");
+
+    session->RibbonDefinitionfileLoad("creo2urdf.rbn");
+
+    return (0);
+}
+
+/*====================================================================*\
+FUNCTION : user_terminate()
+PURPOSE  : To handle any termination actions
+\*====================================================================*/
+extern "C" void user_terminate()
+{
+}
